+++ spimdisasm/mips/MipsFileBase.py
@@ -29,7 +29,7 @@ class FileBase(common.ElementBase):
         self.symbolsVRams: set[int] = set()
         "addresses of symbols in this section"
 
-        self.stringEncoding: str = "ASCII"
+        self.stringEncoding: str = common.GlobalConfig.DATA_STRING_ENCODING
 
         self.bytes: bytes = common.Utils.wordsToBytes(self.words)